requirement.txt
requests>=2.31.0
fpdf2>=2.7.8
weasyprint>=60.1
openai>=1.13.0
faiss-cpu>=1.7.4
tqdm>=4.66.0
python-dotenv>=1.0.0    # (optionnel) pour charger la clé API depuis un fichier .env